+++ dojos/20170927-react-native/demo/components/Details.js
@@ -1,0 +1,23 @@
+import React from 'react';
+import {
+  Text,
+  View,
+} from 'react-native';
+
+import PlaceDetails from './PlaceDetails';
+
+ class Details extends React.Component {
+  static navigationOptions = {
+    title: 'Detalhes',
+  };
+  render() {
+    const { data } = this.props.navigation.state.params;
+    return (
+      <View>
+        <PlaceDetails data={data}></PlaceDetails>
+      </View>
+    );
+  }
+}
+
+export { Details };
